+++ lib/main.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:myapp/providers/StationProvider.dart';
-import 'package:myapp/widgets/FoundObject.dart';
+import 'package:myapp/models/FoundObject.dart';
 import 'widgets/search_button.dart';
 import 'search.dart';
 import 'package:provider/provider.dart';
-import 'providers/ObjectProvider.dart';
+import 'providers/ObjectsProvider.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:shared_preferences/shared_preferences.dart';
